refactor(codex): split cold scanner tests
